Refactor message into model
