+++ WebApp/autoreduce_webapp/reduction_variables/utils.py
@@ -17,14 +17,14 @@ import sys
 import chardet
 import importlib.util as imp
 
-from message.job import Message
+from model.message.job import Message
 
 sys.path.append(os.path.join("../", os.path.dirname(os.path.dirname(__file__))))
 os.environ["DJANGO_SETTINGS_MODULE"] = "autoreduce_webapp.settings"
 
 # pylint:disable=wrong-import-position
 from autoreduce_webapp.icat_communication import ICATCommunication
-from autoreduce_webapp.settings import ACTIVEMQ, REDUCTION_DIRECTORY, FACILITY
+from autoreduce_webapp.settings import REDUCTION_DIRECTORY, FACILITY
 from reduction_variables.models import InstrumentVariable, RunVariable
 from reduction_viewer.models import ReductionRun, Notification
 from reduction_viewer.utils import InstrumentUtils, StatusUtils, ReductionRunUtils
