+++ monitors/tests/test_run_detection.py
@@ -7,7 +7,7 @@ import csv
 from filelock import FileLock
 from mock import (Mock, patch, call)
 
-from message.job import Message
+from model.message.job import Message
 from monitors.settings import (CYCLE_FOLDER, LAST_RUNS_CSV)
 import monitors.run_detection as eorm
 from monitors.run_detection import (InstrumentMonitor,
