+++ queue_processors/autoreduction_processor/post_process_admin.py
@@ -29,7 +29,7 @@ import importlib.util as imp
 from sentry_sdk import init
 
 # pylint:disable=no-name-in-module,import-error
-from message.job import Message
+from model.message.job import Message
 from paths.path_manipulation import append_path
 from queue_processors.autoreduction_processor.settings import MISC
 from queue_processors.autoreduction_processor.autoreduction_logging_setup import logger
