+++ queue_processors/autoreduction_processor/tests/test_post_process_admin.py
@@ -16,7 +16,7 @@ import json
 from tempfile import mkdtemp, NamedTemporaryFile
 from mock import patch, call, Mock
 
-from message.job import Message
+from model.message.job import Message
 from paths.path_manipulation import append_path
 from utils.settings import ACTIVEMQ_SETTINGS
 from utils.project.structure import get_project_root
