+++ utils/clients/tests/test_queue_client.py
@@ -11,7 +11,7 @@ import unittest
 
 from mock import patch, call
 
-from message.job import Message
+from model.message.job import Message
 from utils.clients.connection_exception import ConnectionException
 from utils.clients.queue_client import QueueClient
 from utils.clients.settings.client_settings_factory import ClientSettingsFactory
